requirements/base.txt
Django>2.2,<2.3
psycopg2==2.8.3
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
django-cors-headers==3.1.0
django-filter==2.2.0
Pillow==6.2.0
google-cloud-speech==1.3.2